Share logrotate option list between definition and global configuration
We now use the constants in the CookbookLogrotate module to validate
the global configuration options and the logrotate_app options. This
ensures that the options available to each stay in sync.  The
logrotate_app template is now more dynamically driven, ensuring that
the template does not need to be updated every time an option is
added.