Quick fix for our problem.
We have an scenario were we bake our server images in a bakery and
run some final recipes in 'run time'. We do not have access to any
rpm repository in 'run time' so we can not install any package in that
step. The fact that the cookbook always include the default recipe in
the lorgorate_app definition make this task impossible.