Wont call default recipe in logrotate_app definition
This change is important in our use case, but it may cause some
problems to existing users of the cookbook. The problem calling
logrotate::default inside the logrotate_app definition is that if
you can't reach the rpm repotitory in that moment you can't configure
the logrotate. We are baking our AMIs and deploying them in a VPC
without access to any repository, some configurations should be applied
inside the VPC but we can not configure any repository or install
any package in that moment.